feat(Image): Render children appropriately
Previously you could technically pass `children` but they'd be passed to the
<img>, which is not technically valid markup.

This also brings this more inline with the Input component. An HTMLInput
doesn't allow children, but a SUI Input is a wrapper around an HTMLInput
so it technically could. Since there could be other elements inside a
`.ui.image` wrapper this gives the user more control over them.