docs(StepExampleGroups): remove object destructuring
Although this isn't a bug and isn't necessary, I'd like to propose reworking the Step examples that use object destructuring. All of the other examples that I've seen in the docs that have subcomponents don't use destructuring. Because of that, it is easy to get mixed up on the destructuring when looking at tags without the base component name in there. It looks like maybe there's a Group tag, etc. Once it is understood then developers can do the destructuring and be aware, but this creates a sort of lowest common denominator for all styles of programming and keeps it consistent with what I've observed in the rest of the examples.

If this is agreeable I can go through and change the other Step examples that have the same destructuring in them.