Allow "right" to be passed as `actionPosition` on Input element
`actionPosition` and `iconPosition` both default to "right", but prop-types for both warn when "right" is passed in. Came up in the following case:

```jsx
<Form.Input label="Some label" actionPosition="right">
  <input />
  {flag && <Button content="Sometimes Visible" />}
  <Button content="Always Visible" />
</Form.Input>
```

Simply letting the prop default to "right" doesn't work as expected: the "Always Visible" button appears separated from the input and "Sometimes Visible" button is always hidden.