fix(release): keep dropdown open when closeOnChange is set to false
When options prop is used, clicking on the dropdown item will trigger handleItemClick which will utilize closeOnChange properly.
When options prop is undefined, clicking on the dropdown item will trigger handleClick instead which will use toggle instead of closeOnChange.
Thus, replaced this.close(e) in toggle with this.closeOnChange(e) to keep dropdown open when closeOnChange is set to false.

issue: Dropdown: ignores closeOnChange if it has children and isn't a search #2041