Fix Issue #2388
Details: The issue was related to portals and there management of outside-clicks, where Portal
support dismissing the Portal children component once a click outside the Portal is recongnize. The
issue can be reproduce on any element (Not just the Select component) by clicking on element that will be deleted
from the DOM, In this case the click on the icon-delete button (x) was doing some internal logic of
the Select component and then the icon-delete will be removed from the DOM, after that the click will be bubbled to the Portal
and the portal will decide if the click is in/out the Portal to handle auto-close feature.

The Portal will search for icon-delete and see if it in the this.rootNode of the Portal, and because has already deleted the removal
icon from the DOM tree it will not find it in the rootNode and it will decide that it's was a click outside the Portal! Also, the event.target
won't keep a full reference to the top body element in case we want to search the icon-delete parents! so the way is to use event.path / event.composedPath
to have the full snapshot of the bubbled event