Use string enums rather than string unions to define types
Specifying the value of `color` on a `Progress` component yields a TypeScript error, even if passing in a defined string constant (such as `green`). Is there any reason the union types such as `SemanticCOLORS` and `SemanticDIRECTIONALTRANSITIONS` aren't defined as string enums?

I've included a proposed fix for the `SemanticCOLORS` type. By turning it into an enum, you get an inferred type definition as well as string values that can be used at runtime. Happy to submit a complete PR transforming all string unions into enums if this seems to be interesting to the maintainers.