Bug: Put contentRef in state to let EventStack track changes
contentRef is passed to EventStack to subscribe for mouse events.
Passing a `contentRef` created using `React.createRef()` does not allow
EventStack to keep track of contentRef changes (since only
contentRef.current changes).
Ultimately when PortalInner unmount contentRef.current is set to undefined
and EventStack `unsubscribe` handlers with `undefined` as
a target instead of unsubscribing with real content.
This leads to the event-stack Map keeping a reference to the real contentRef
thus leaking all dom tree related to the node related to contentRef.

Putting contentRef in the state and passing the element allow for
EventStack to detect change in props and execute the componentDidUpdate
routine which unsubscribe from the previous element on unmount.